backend/rag_agent/requirements.txt
 fastapi
 openai
 qdrant-client
 python-dotenv
 uvicorn
 pydantic